requirements.txt
fastapi
uvicorn
pydantic[email]
pymongo
python-dotenv
requests
razorpay